refactor: extract dev-tools router (PR-5, Phase 1)
Pure move of 7 dev/debug/test endpoints from server/routes.ts into
server/routes/devTools.ts, following the full-path router convention
(PRs 2-4). Mounted app.use(devToolsRouter) at the original position of
the first endpoint (GET /api/debug/data-load) to preserve registration
order.

Endpoints moved verbatim (handlers, log lines, status codes, response
shapes, per-route middleware):
- GET  /api/debug/data-load          (requireClerkUser)
- GET  /api/test-data                (requireClerkUser)
- GET  /api/validate-types           (requireClerkUser)
- GET  /api/debug/game/:gameId/revenue        (no auth)
- GET  /api/dev/markets-config       (requireClerkUser)
- POST /api/dev/markets-config       (requireClerkUser)
- POST /api/game/:gameId/test/streaming-decay (requireClerkUser)

Two TODO comments added (no behavior change) flagging the no-auth debug
revenue endpoint and the non-admin markets-config write endpoint.

Removed fs/path imports from routes.ts (now unused after the move);
gameDataLoader stays (still used elsewhere). Route-manifest snapshot
unchanged. npm run check clean; 545 tests pass.
